+++ lib/Models/CsvCatalogItem.ts
@@ -17,6 +17,7 @@ import AutoRefreshingMixin from "../ModelMixins/AutoRefreshingMixin";
 import isDefined from "../Core/isDefined";
 import DiscretelyTimeVaryingMixin from "../ModelMixins/DiscretelyTimeVaryingMixin";
 import { BaseModel } from "./Model";
+import ExportableData from "./ExportableData";
 
 // Types of CSVs:
 // - Points - Latitude and longitude columns or address
@@ -31,19 +32,21 @@ import { BaseModel } from "./Model";
 
 const automaticTableStylesStratumName = "automaticTableStyles";
 
-export default class CsvCatalogItem extends AsyncChartableMixin(
-  TableMixin(
-    // Since both TableMixin & DiscretelyTimeVaryingMixin defines
-    // `chartItems`, the order of mixing in is important here
-    DiscretelyTimeVaryingMixin(
-      AsyncMappableMixin(
-        AutoRefreshingMixin(
-          UrlMixin(CatalogMemberMixin(CreateModel(CsvCatalogItemTraits)))
+export default class CsvCatalogItem
+  extends AsyncChartableMixin(
+    TableMixin(
+      // Since both TableMixin & DiscretelyTimeVaryingMixin defines
+      // `chartItems`, the order of mixing in is important here
+      DiscretelyTimeVaryingMixin(
+        AsyncMappableMixin(
+          AutoRefreshingMixin(
+            UrlMixin(CatalogMemberMixin(CreateModel(CsvCatalogItemTraits)))
+          )
         )
       )
     )
   )
-) {
+  implements ExportableData {
   static get type() {
     return "csv";
   }
@@ -73,6 +76,30 @@ export default class CsvCatalogItem extends AsyncChartableMixin(
   @computed
   get hasLocalData(): boolean {
     return isDefined(this._csvFile);
+  }
+
+  async exportData() {
+    if (isDefined(this._csvFile)) {
+      return {
+        name: (this.name || this.uniqueId)!,
+        file: this._csvFile
+      };
+    }
+    if (isDefined(this.csvString)) {
+      return {
+        name: (this.name || this.uniqueId)!,
+        file: new Blob([this.csvString])
+      };
+    }
+
+    if (isDefined(this.url)) {
+      return this.url;
+    }
+
+    throw new TerriaError({
+      sender: this,
+      message: "No data available to download."
+    });
   }
 
   @computed